udemy_langgraph/reflection_agent/pyproject.toml
[project]
name = "reflection-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "langchain-core>=0.3.34",
    "langchain-openai>=0.3.4",
    "langgraph>=0.2.70",
    "python-dotenv>=1.0.1",
]
